experimental/keyphrase_extraction_skill/requirements.txt
azure-identity==1.21.0
azure-mgmt-core==1.5.0
azure-mgmt-search==9.1.0
azure-mgmt-storage==22.2.0
azure-search-documents==11.5.2
dotenv==0.9.9